+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed:
   * Native Linux saves were not detected in some Flatpak roots.
     ([Contributed by madscientist16](https://github.com/mtkennerly/ludusavi/pull/556))
+  * For Lutris roots, some GOG install folders were not properly detected.
 
 ## v0.30.0 (2025-11-09)